+++ @types/enums.ts
@@ -1,0 +1,7 @@
+export enum ONESIGNAL_ERROR_TYPE {
+  INVALID_EXTERNAL_USER_IDS = "invalid_external_user_ids"
+}
+
+export enum NotificationsType {
+  CHAT_MESSAGE = "CHAT_MESSAGE"
+}